+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 All notable changes to `Enumhancer` will be documented in this file
 
+## 1.7.0 - 2022-06-06
+- When using [Comparison](docs/comparison.md), you can now assert with `is` or `isNot`
+
 ## 1.6.0 - 2022-06-04
 - added Eloquent Casting support for basic enumerations.
 
 ## 1.5.0 - 2022-05-31
-- added `Extractor` to extract enums from a string mentioned by value
+- added [Extractor](docs/extractor.md) to extract enums from a string mentioned by value
 - some documentation repairs
 
 ## 1.4.1 - 2022-03-04